Genericize fs
