+++ snapfaas/bins/multivm/main.rs
@@ -12,7 +12,7 @@
 use clap::{App, Arg};
 use log::warn;
 use snapfaas::resource_manager::ResourceManager;
-use snapfaas::{sched, labeled_fs};
+use snapfaas::{sched, fs::lmdb::DBENV};
 use snapfaas::worker::Worker;
 
 use std::net::{SocketAddr, TcpStream};
@@ -76,7 +76,7 @@ fn main() {
     let pool_size = manager.total_mem_in_mb() / 128;
     let pool = threadpool::ThreadPool::new(pool_size);
     let manager = Arc::new(Mutex::new(manager));
-    let db = &*labeled_fs::DBENV;
+    let db = &*DBENV;
     for i in 0..pool_size as u32 {
         let sched_addr_dup = sched_addr.clone();
         let manager_dup = Arc::clone(&manager);
